pyproject.toml
[tool.poetry]
name = "nomansland-dev"
version = "0.1.0"
description = ""
authors = ["Saranya Balasubramanian <saranya.balasubramanian@oeaw.ac.at>"]
readme = "README.md"
packages = [{include = "apis_ontology"}]

[tool.poetry.dependencies]
python = "^3.11"
apis-core-rdf = "0.36.1"
apis-acdhch-default-settings = "2.0.0"
psycopg2 = "^2.9.10"
django-reversion = "^5.0.12"
apis-highlighter-ng = "^0.5.0"
django-acdhch-functions = "^0.1.3"
django-extensions = "^3.2.3"
pandas = "^2.2.2"
tqdm = "^4.67.1"
apis-bibsonomy = "^0.13.1"
django-interval = "^0.2.5"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
